Removed AwesomeNetworkHelper and created shared instance of AwesomeNetwork
