+++ Example/AwesomeNetwork/AppDelegate.swift
@@ -17,7 +17,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-        AwesomeNetwork.shared?.startNetworkStateNotifier()
+
+        AwesomeNetwork.shared.startNetworkStateNotifier()
+        
         return true
     }
 
@@ -29,11 +31,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationDidEnterBackground(_ application: UIApplication) {
         // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
         // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
-        AwesomeNetwork.shared?.stopNetworkStateNotifier()
+        
+        AwesomeNetwork.shared.stopNetworkStateNotifier()
     }
 
     func applicationWillEnterForeground(_ application: UIApplication) {
         // Called as part of the transition from the background to the inactive state; here you can undo many of the changes made on entering the background.
+        AwesomeNetwork.shared.startNetworkStateNotifier()
     }
 
     func applicationDidBecomeActive(_ application: UIApplication) {
@@ -42,6 +46,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationWillTerminate(_ application: UIApplication) {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
+        
+        AwesomeNetwork.shared.stopNetworkStateNotifier()
     }
